Add code to work with socket
